works for non-mod

+++ Interpolator/Point.java
@@ -1,9 +1,9 @@
 public class Point {
 
-	int x;
-	int y;
+	double x;
+	double y;
 
-	public Point(int x, int y) {
+	public Point(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
